requirements.txt
streamlit
cryptography
numpy
google-generativeai
pandas
PyPDF2
